worked on issues

+++ app/Http/Controllers/ListingController.php
@@ -240,30 +240,55 @@ class ListingController extends Controller
      */
     public function update(Request $request, $postId)
     {
-        $this->googleService->updatePost($request->all(), $postId);
+        $data = $this->googleService->updatePost($request->all(), $postId);
 
-        if ($request->database) {
-            $listing = Listing::where('product_id', $request->database)
-                ->first();
+        if ($data->getStatusCode() != 500) {
+            if ($request->database) {
+                $listing = Listing::where('product_id', $request->database)
+                    ->first();
 
-            $additionalInfo = UserListingInfo::where('title', $listing->title)
-                ->where('image', $listing->images)
-                ->first();
+                $additionalInfo = UserListingInfo::where('title', $listing->title)
+                    ->where('image', $listing->images)
+                    ->first();
 
-            $listing->delete();
+                $listing->delete();
 
-            if ($additionalInfo) {
-                $additionalInfo->update([
-                    'status' => 1,
-                    'approved_by' => auth()->user()->id,
-                    'approved_at' => now()
-                ]);
+                if ($additionalInfo) {
+                    $additionalInfo->update([
+                        'status' => 1,
+                        'approved_by' => auth()->user()->id,
+                        'approved_at' => now()
+                    ]);
+                }
+
+                return redirect()->route('database-listing.index', ['status' => 0, 'startIndex' => 1, 'category' => '', 'user' => 'all']);
             }
 
-            return redirect()->route('database-listing.index', ['status' => 0, 'startIndex' => 1, 'category' => '', 'user' => 'all']);
-        }
+            if (request()->edit == 'true') {
+                $listing = Listing::where('product_id', $request->product_id)
+                    ->first();
 
-        session()->flash('success', 'Post updated successfully');
+                $additionalInfo = UserListingInfo::where('title', $listing->title)
+                    ->where('image', $listing->images)
+                    ->first();
+
+                $listing->delete();
+
+                if ($additionalInfo) {
+                    $additionalInfo->update([
+                        'status' => 1,
+                        'approved_by' => auth()->user()->id,
+                        'approved_at' => now()
+                    ]);
+                }
+            }
+
+            session()->flash('success', 'Post updated successfully');
+
+            return redirect()->route('inventory.index', ['startIndex' => 1, 'category' => 'Product']);
+        }
+        
+        session()->flash('success', 'Please Authenticate with Google');
 
         return redirect()->route('inventory.index', ['startIndex' => 1, 'category' => 'Product']);
     }
